Feat : APItype 코드 수정
