+++ src/Network/APIType.js
@@ -1,7 +1,7 @@
 export function url(path) {
   // const version = 'v1'; // process.env.REACT_APP_API_VERSION
 
-  return `https://api-alpha.42world.kr/${path}`;
+  return `https://api-alpha.42world.kr${path}`;
 }
 
 // 없어도 되는 코드
